ACPI / LPSS: add support of shared clock
The DMA controller in Lynxpoint is using the LPSS root clock as a functional
clock to work properly. Proposed patch adds support of the shared clock. It
also avoids hardcoding the name of the DMA device in clk-lpt.c and the name of
the root clock in acpi_lpss.c.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>