i2c-designware: always clear interrupts before enabling them
The controller might still have state of the previous transfer in its raw
interrupt register so make sure that state is cleared before starting the
next transfer.

For example if the i2c bus is put into low power state by an ACPI method it
might pull the SDA line low. Once the bus is put to full power state it
pulls the SDA line back to high and this looks like STOP condition from the
controller point-of-view. So it sets STOP detected status bit in its raw
interrupt status register causing the driver to fail the next transfer.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>