i2c-designware: prevent signal from interrupting I2C transfer
If the process receives a signal while it is waiting for I2C transfer to
complete an error is returned to the caller and the transfer is cancelled.
Better is to let the I2C transfer to complete before the signal is handled.
We still use wait_for_completion_timeout() to prevent any hangs on the bus.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>