ACPI / LPSS: add support for Intel BayTrail
Intel BayTrail has almost the same Low Power Subsystem than Lynxpoint with
few differences. Peripherals are clocked with different speeds (typically
lower) and the clock is not always gated. To support this we add
possibility to share a common fixed rate clock and make clock gating
optional.

Backport of 3.11 commit f627217064dbef1eef53ceb01edb9c94203991e0. This
patch was disabled by removing the Baytrail IDs from the LPSS list because
the mainline Baytrail enabeling collides with the ABT Baytrail enabeling.
This will have to be resolved in the future and at that time the rest of
this patch should be backported. The reason for the partial backport is
that future patches depend on this one.

Issue: APDEV-1386
Change-Id: I4c5046b2072636625e3f63890e61f805ca74ffcd
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Acked-by: Mike Turquette <mturquette@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Reviewed-on: https://android.intel.com/160009
Reviewed-by: Mossberg, Johan <johan.mossberg@intel.com>
Tested-by: Mossberg, Johan <johan.mossberg@intel.com>
Tested-by: sys_abtbuild <sys_abtbuild@intel.com>
Reviewed-by: Holmberg, Hans <hans.holmberg@intel.com>
Reviewed-by: Odeberg, Hans <hans.odeberg@intel.com>