ACPI / LPSS: mask the UART TX completion interrupt
Intel LPSS provides an extra TX byte counter and an extra TX
completion interrupt for some of its bus controllers.  However,
there is no use for the extra UART interrupt and it has to be
masked out during initialization.

Otherwise, if the firmware does not mask the interrupt and
the driver does not clear it, it may cause an interrupt flood
freezing the board to happen.

Add code masking that problematic interrupt to the ACPI LPSS driver.

Backport of 3.11 commit 06d8641504726322fca54400bbac982bd44f9a27

Issue: APDEV-1386
Change-Id: Id81e3675e40190bb0b645ff20cffbdbdb2782341
[rjw: Changelog]
Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Reviewed-on: https://android.intel.com/160010
Reviewed-by: Mossberg, Johan <johan.mossberg@intel.com>
Tested-by: Mossberg, Johan <johan.mossberg@intel.com>
Tested-by: sys_abtbuild <sys_abtbuild@intel.com>
Reviewed-by: Nordell, Joakim <joakim.nordell@intel.com>
Reviewed-by: Odeberg, Hans <hans.odeberg@intel.com>