ACPI / LPSS: fix UART Auto Flow Control
There is an additional bit in the GENERAL register on newer
silicon that needs to be set or UART's RTS pin fails to
reflect the flow control settings in the Modem Control
Register.

This will fix an issue where the RTS pin of the UART stays
always at 1.8V, regardless of the register settings.

Backport of 3.13 commit 088f1fd267c7f43b5d87850a0fa0c7e851ecae97

Issue: APDEV-1386
Change-Id: I635b8bcb15abea6aae09ec99c2a73f7e2ef72d40
Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Reviewed-on: https://android.intel.com/160012
Reviewed-by: Mossberg, Johan <johan.mossberg@intel.com>
Tested-by: Mossberg, Johan <johan.mossberg@intel.com>
Tested-by: sys_abtbuild <sys_abtbuild@intel.com>
Reviewed-by: Nordell, Joakim <joakim.nordell@intel.com>
Reviewed-by: Odeberg, Hans <hans.odeberg@intel.com>